requirements.txt
# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0

# 数据库
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# 任务队列
celery>=5.3.6
redis>=5.0.1
kombu>=5.3.5

# HTTP 客户端
httpx>=0.26.0
aiohttp>=3.9.1

# NLP & ML
openai>=1.12.0
jieba>=0.42.1
snownlp>=0.12.3

# 数据处理
pandas>=2.2.0
numpy>=1.26.3
pydantic>=2.5.3

# 工具
python-dotenv>=1.0.0
loguru>=0.7.2
schedule>=1.2.0

# RSS/新闻
feedparser>=6.0.10
newspaper3k>=0.2.8
lxml>=5.1.0
html2text>=2020.1.16

# 监控
prometheus-client>=0.19.0
